Move endpoint to individual files for better seperation of endpoints.
