+++ backend/routers/courses.py
@@ -1,0 +1,30 @@
+from fastapi import APIRouter, HTTPException, status
+
+router = APIRouter(
+    prefix="/courses",
+    tags=["Courses"],
+)
+
+@router.post("/courses")
+async def create_course():
+    """Placeholder: This endpoint will create a new course."""
+    # Logic will be protected and require an authenticated user (via JWT).
+    return {"message": "Create course endpoint placeholder."}
+
+@router.get("/courses")
+async def get_all_courses():
+    """Placeholder: This endpoint will retrieve all courses for the logged-in user."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": "Get all courses endpoint placeholder."}
+
+@router.get("/courses/{course_id}")
+async def get_course(course_id: str):
+    """Placeholder: This endpoint will retrieve a specific course by its ID."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": f"Get course {course_id} endpoint placeholder."}
+
+@router.put("/courses/{course_id}")
+async def update_course(course_id: str):
+    """Placeholder: This endpoint will update a specific course by its ID."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": f"Update course {course_id} endpoint placeholder."}
