+++ backend/routers/tasks.py
@@ -1,0 +1,37 @@
+from fastapi import APIRouter, HTTPException, status
+import pydantic
+
+router = APIRouter(
+    prefix="/tasks",
+    tags=["Tasks"],
+)
+
+@router.get("/", tags=["Tasks"])
+async def get_all_tasks():
+    """Placeholder: This endpoint will retrieve all tasks for the logged-in user."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": "Get all tasks endpoint placeholder."}
+
+@router.post("/", tags=["Tasks"])
+async def create_task():
+    """Placeholder: This endpoint will create a new task."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": "Create task endpoint placeholder."}
+
+@router.get("/{task_id}", tags=["Tasks"])
+async def get_task(task_id: str):
+    """Placeholder: This endpoint will retrieve a specific task by its ID."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": f"Get task {task_id} endpoint placeholder."}
+
+@router.put("/{task_id}", tags=["Tasks"])
+async def update_task(task_id: str):    
+    """Placeholder: This endpoint will update a specific task by its ID."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": f"Update task {task_id} endpoint placeholder."}
+
+@router.delete("/{task_id}", tags=["Tasks"])
+async def delete_task(task_id: str):
+    """Placeholder: This endpoint will delete a specific task by its ID."""
+    # Logic will be protected and require an authenticated user.
+    return {"message": f"Delete task {task_id} endpoint placeholder."}
